requirements.txt
2captcha-python
aptos-sdk_0.6.3
